Move inscription type to utils
